// src/api/generated/models/diagnosisResponseSchema.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * AI Agro API
 * API для AI Agro: регистрация, авторизация, создание чата и диагностика растений.
 * OpenAPI spec version: 1.0.0
 */
import type { DiagnosisOutSchema } from './diagnosisOutSchema';

export interface DiagnosisResponseSchema {
  diagnosis?: DiagnosisOutSchema | null;
  answer: string;
  should_save?: boolean;
}